requirements.txt
streamlit
PyMuPDF
python-dotenv
tqdm
google-generativeai
